user_about_in[f][t]ro. md
TLDR of the (I think it is, but I also consciously lie and sabotage myself, from day one on this louy in the "user_about_in[f][t]ro. md" file:

As a kid, I was a super cirous of the surroundings and obnoxiously agile, nosey little bastard. Were, was, am and always will be: driven by the thirst for knowledge and understanding of all things, both physical, as well not, but steixrly limited to scientifically prove[n]/[able] topics; zero nonsense tolerance, extremely allergic to bulshit; especially moddern "youtube-is-a-great-and-trustwordy-accurate-&-ONE-HUNDRED-PERCENT-authentic-source-for-NikolaTesla, JFK, theMoon, earthCurveture type of things and topics! 
School-book eligibility enforcer! If it wasnt, or will never be written in the the big-phat-books